+++ packages/ethereum-contracts/test/testenv-ctl.sh
@@ -3,21 +3,19 @@
 # make sure that if any step fails, the script fails
 set -xe
 
-TESTENV_MNEMONIC="candy maple cake sugar pudding cream honey rich smooth crumble sweet treat"
-
 CMD=$1
 
-start_ganache() {
-    ganache-cli --networkId 4447 --port 47545 --mnemonic "$TESTENV_MNEMONIC"
+start_hardhat_node() {
+    npx hardhat node --port 47545
 }
 
-kill_ganache() {
-    pkill -f "ganache-cli --networkId 4447" || true
+kill_hardhat_node() {
+    pkill -f "hardhat node --port 47545" || true
 }
 
 if [ "$CMD" == "start" ];then
-    kill_ganache
-    start_ganache
+    kill_hardhat_node
+    start_hardhat_node
 elif [ "$CMD" == "stop" ];then
-    kill_ganache
+    kill_hardhat_node
 fi